requirements.txt
flask
gspread
google-auth
google-auth-oauthlib
requests
beautifulsoup4
gunicorn
python-dotenv
pytest
pytest-mock
